Se han comentado todas las funciones del archivo 'discPayBack.js'

+++ js-scripts/discPayBack.js
@@ -1,4 +1,13 @@
-define(['jquery'], function($) {
+/**
+ * File
+ * Este archivo contiene todas las funciones que realizan el cálculo de 'Discounted Payback Period'
+ * 
+ */
+//Está linea se compone de un array donde colocaremos entre comillas '' la biblioteca
+//que será usada por nuestras funciones y una función callback que servirá para relacionar el 
+//nombre de la biblioteca del array con un simbolo, en este caso el '$' ya que jquery utiliza por defecto
+//este símbolo para hacer uso de sus funciones
+define(['jquery'], function($) { 
     var methods = {};
 
     methods.displayPeriodosTab1 = displayPeriodosTab1;
@@ -12,8 +21,8 @@ define(['jquery'], function($) {
 });
 
 /**
- * @brief muestra en la GUI una tabla de 4*N celdas, siendo N el número de periodos
- * @param numero_periodos es un número entero positivo que tiene significado por su propio nombre
+ * @brief muestra en la tabla del tab 1 una tabla de 4*N celdas, siendo N el número de periodos/filas
+ * @param {number} numero_periodos - es un número entero positivo que tiene significado por su propio nombre
  */
 function displayPeriodosTab1(numero_periodos){//Despliega filas de periodos con 3 columnas 
     var template;
@@ -36,9 +45,9 @@ function displayPeriodosTab1(numero_periodos){//Despliega filas de periodos con 
 }
 
 /**
- * @brief muestra en la columna 'Cumulative Cash Flow' de la GUI el contenido del array cumCashFlow[]
- * @param numero_periodos es un número entero positivo que tiene significado por su propio nombre
- * @param cumCashFlow es un vector de números reales(+/-) que guarda la los valores de Cumulative Cash Flow de 1-n periodos
+ * @brief muestra en la columna 'Cumulative Cash Flow' de la tabla del tab 1 los datos correspondientes
+ * @param {number} numero_periodos - es un número entero positivo que tiene significado por su propio nombre
+ * @param {Array<number>} cumCashFlow - es un vector de números reales(+/-) que guarda la los valores de Cumulative Cash Flow de [1,N] periodos
  */
 function displayCumCashFlow(numero_periodos, cumCashFlow){
     for(var i = 1; i <= numero_periodos; i++){
@@ -46,9 +55,9 @@ function displayCumCashFlow(numero_periodos, cumCashFlow){
     }
 }
 /**
- * @brief obtiene los datos de la columna de 'outflows' de la GUI y lo convierte a un array
- * @param numero_periodos es un número entero positivo que tiene significado por su propio nombre
- * @return un arreglo de números reales(+/-) que contiene a los outflows de 1-n periodos
+ * @brief obtiene los datos de la columna de 'outflows' de la tabla del tab 1
+ * @param {number} numero_periodos - es un número entero positivo que tiene significado por su propio nombre
+ * @returns {Array<number>} un arreglo de números reales(+/-) que contiene a los 'outflows' de [1,N] periodos
  */
 function getOutflows(numero_periodos){
     var data = [];
@@ -58,9 +67,9 @@ function getOutflows(numero_periodos){
         return data;
 }
 /**
- * @brief obtiene los datos de la columna de 'inflows' de la GUI y los guarda en un array
- * @param numero_periodos es un número entero positivo que tiene significado por su propio nombre
- * @return un arreglo de números reales que contiene los outflows de 1-n periodos
+ * @brief obtiene los datos de la columna de 'inflows' de la tabla del tab 1
+ * @param {number} numero_periodos - es un número entero positivo que tiene significado por su propio nombre
+ * @returns {Array<number>} un arreglo de números reales que contiene los outflows de [1,N] periodos
  */
 function getInflows(numero_periodos){
     var data = [];
@@ -72,9 +81,10 @@ function getInflows(numero_periodos){
 }
 /**
  * @brief calcula el 'Discounted Cash Flow'
- * @param netCashFlow[] array de números reales(+/-) que guarda los valores del net cash flow de 1-n periodos
- * @param npv[] array que guarda los valores del nvp de 1-n periodos
- * @param numero_periodos es un número entero positivo que tiene significado por su propio nombre
+ * @param {Array<number>} netCashFlow - array de números reales(+/-) que guarda los valores del 'net cash flow' de [1,N] periodos
+ * @param {Array<number>} npv - array que guarda los valores del nvp de [1,N] periodos
+ * @param {number} numero_periodos -  es un número entero positivo que tiene significado por su propio nombre
+ * @returns {Array<number>} un array que contiene los registros del 'Discounted Cash Flow'
  */
 function calculateDiscCashFlow(netCashFlow, npv, numero_periodos){
     var discCashFlow = [];
@@ -85,13 +95,14 @@ function calculateDiscCashFlow(netCashFlow, npv, numero_periodos){
 }
 /**
  * @brief calcula el 'Cumulative Cash Flow'
- * @param principal es un número entero positivo que representa la inversión inicial
- * @param discCashFlow es un array de números reales(+/-) que guarda los valores del discounted cash flow de 1-n periodos 
- * @param periodos es un número entero positivo que tiene significado por su propio nombre
+ * @param {number} principal - es un número entero positivo que representa la inversión inicial
+ * @param {Array<number>} discCashFlow - es un array de números reales(+/-) que guarda los valores del 'discounted cash flow' de [1,N] periodos 
+ * @param {number} periodos - es un número entero positivo que tiene significado por su propio nombre
+ * @returns {Array<number>} un array que contiene los registros del 'Cumulative Cash Flow'
  */
 function calculateCumCashFlow(principal, discCashFlow, periodos){
-    //Convertimos el argumento Principal en INT
-    principal = parseInt(principal);
+    //Convertimos el argumento Principal en Floar
+    principal = parseFloat(principal);
 
     var cumCashFlow = [];
     cumCashFlow[1] = principal + discCashFlow[1];
